Add Voice Picker
Previously we would just cycle through 3 voices I had tried. The
default values is still those voices (for now), but the user can now also
add their own voices, along with customizing the Role Play and Speaking
Style options provided by azure.

In the future I may get it so that you can select specific voices for
specific cards / books etc

+++ backend/generateSentences_test.go
@@ -32,7 +32,7 @@ func TestGenerate(t *testing.T) {
 	os.Remove(tempDb)
 	defer os.Remove(tempDb)
 	myBackend := createBackend(tempDb)
-	bookIds := []int64{1}
+	bookIds := []int{1}
 	_, err := myBackend.BookLibrary.AddBook("张天翼", "秃秃大王", "fake.jpg", "testdata/example_book.txt")
 	myBackend.Generator.GenerateSentenceTable()
 	if err != nil {
